Update `HEVC` function

+++ src/Sophia_Script_for_Windows_10_PowerShell_7/Module/Sophia.psm1
@@ -534,7 +534,7 @@ function ScheduledTasks
 	function DisableButton
 	{
 		Write-Information -MessageData "" -InformationAction Continue
-		# Extract the localized "Please wait..." string from shell32.dll
+		# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 		Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 		[void]$Window.Close()
@@ -546,7 +546,7 @@ function ScheduledTasks
 	function EnableButton
 	{
 		Write-Information -MessageData "" -InformationAction Continue
-		# Extract the localized "Please wait..." string from shell32.dll
+		# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 		Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 		[void]$Window.Close()
@@ -599,7 +599,7 @@ function ScheduledTasks
 		"Enable"
 		{
 			$State           = "Disabled"
-			# Extract the localized "Enable" string from shell32.dll
+			# Extract the localized "Enable" string from %SystemRoot%\System32\shell32.dll
 			$ButtonContent   = [WinAPI.GetStrings]::GetString(51472)
 			$ButtonAdd_Click = {EnableButton}
 		}
@@ -612,7 +612,7 @@ function ScheduledTasks
 	}
 
 	Write-Information -MessageData "" -InformationAction Continue
-	# Extract the localized "Please wait..." string from shell32.dll
+	# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 	Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 	# Getting list of all scheduled tasks according to the conditions
@@ -2890,7 +2890,7 @@ function UnpinTaskbarShortcuts
 		$Shortcuts
 	)
 
-	# Extract the localized "Unpin from taskbar" string from shell32.dll
+	# Extract the localized "Unpin from taskbar" string from %SystemRoot%\System32\shell32.dll
 	$LocalizedString = [WinAPI.GetStrings]::GetString(5387)
 
 	foreach ($Shortcut in $Shortcuts)
@@ -2904,7 +2904,7 @@ function UnpinTaskbarShortcuts
 					# Call the shortcut context menu item
 					$Shell = (New-Object -ComObject Shell.Application).NameSpace("$env:AppData\Microsoft\Internet Explorer\Quick Launch\User Pinned\TaskBar")
 					$Shortcut = $Shell.ParseName("Microsoft Edge.lnk")
-					# Extract the localized "Unpin from taskbar" string from shell32.dll
+					# Extract the localized "Unpin from taskbar" string from %SystemRoot%\System32\shell32.dll
 					$Shortcut.Verbs() | Where-Object -FilterScript {$_.Name -eq $LocalizedString} | ForEach-Object -Process {$_.DoIt()}
 				}
 			}
@@ -2912,7 +2912,7 @@ function UnpinTaskbarShortcuts
 			{
 				if ((New-Object -ComObject Shell.Application).NameSpace("shell:::{4234d49b-0245-4df3-b780-3893943456e1}").Items() | Where-Object -FilterScript {$_.Name -eq "Microsoft Store"})
 				{
-					# Extract the localized "Unpin from taskbar" string from shell32.dll
+					# Extract the localized "Unpin from taskbar" string from %SystemRoot%\System32\shell32.dll
 					((New-Object -ComObject Shell.Application).NameSpace("shell:::{4234d49b-0245-4df3-b780-3893943456e1}").Items() | Where-Object -FilterScript {
 						$_.Name -eq "Microsoft Store"
 					}).Verbs() | Where-Object -FilterScript {$_.Name -eq $LocalizedString} | ForEach-Object -Process {$_.DoIt()}
@@ -4280,7 +4280,7 @@ function PinToStart
 
 	process
 	{
-		# Extract the localized "Devices and Printers" string from shell32.dll
+		# Extract the localized "Devices and Printers" string from %SystemRoot%\System32\shell32.dll
 		$DevicesPrinters = [WinAPI.GetStrings]::GetString(30493)
 
 		# Checking whether an argument is "DevicesPrinters". The Devices and Printers's AppID attribute can be retrieved only if the shortcut was created
@@ -4550,7 +4550,7 @@ function OneDrive
 			Get-Process -Name explorer | Stop-Process -Force
 
 			Write-Information -MessageData "" -InformationAction Continue
-			# Extract the localized "Please wait..." string from shell32.dll
+			# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 			Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 			Start-Sleep -Seconds 3
@@ -4572,7 +4572,7 @@ function OneDrive
 			Start-Process -FilePath "$env:SystemRoot\explorer.exe"
 
 			Write-Information -MessageData "" -InformationAction Continue
-			# Extract the localized "Please wait..." string from shell32.dll
+			# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 			Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 			Start-Sleep -Seconds 3
@@ -5254,7 +5254,7 @@ function WindowsFeatures
 	function DisableButton
 	{
 		Write-Information -MessageData "" -InformationAction Continue
-		# Extract the localized "Please wait..." string from shell32.dll
+		# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 		Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 		[void]$Window.Close()
@@ -5266,7 +5266,7 @@ function WindowsFeatures
 	function EnableButton
 	{
 		Write-Information -MessageData "" -InformationAction Continue
-		# Extract the localized "Please wait..." string from shell32.dll
+		# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 		Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 		[void]$Window.Close()
@@ -5336,7 +5336,7 @@ function WindowsFeatures
 	}
 
 	Write-Information -MessageData "" -InformationAction Continue
-	# Extract the localized "Please wait..." string from shell32.dll
+	# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 	Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 	# Getting list of all optional features according to the conditions
@@ -5583,7 +5583,7 @@ function WindowsCapabilities
 	function UninstallButton
 	{
 		Write-Information -MessageData "" -InformationAction Continue
-		# Extract the localized "Please wait..." string from shell32.dll
+		# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 		Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 		[void]$Window.Close()
@@ -5597,7 +5597,7 @@ function WindowsCapabilities
 		try
 		{
 			Write-Information -MessageData "" -InformationAction Continue
-			# Extract the localized "Please wait..." string from shell32.dll
+			# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 			Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 			[void]$Window.Close()
@@ -5683,7 +5683,7 @@ function WindowsCapabilities
 	}
 
 	Write-Information -MessageData "" -InformationAction Continue
-	# Extract the localized "Please wait..." string from shell32.dll
+	# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 	Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 	# Getting list of all capabilities according to the conditions
@@ -6189,7 +6189,7 @@ function NetworkAdaptersSavePower
 		)
 		{
 			Write-Information -MessageData "" -InformationAction Continue
-			# Extract the localized "Please wait..." string from shell32.dll
+			# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 			Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 			Start-Sleep -Seconds 2
@@ -6235,7 +6235,7 @@ function NetworkAdaptersSavePower
 		)
 		{
 			Write-Information -MessageData "" -InformationAction Continue
-			# Extract the localized "Please wait..." string from shell32.dll
+			# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 			Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 			Start-Sleep -Seconds 2
@@ -6382,14 +6382,14 @@ function Set-UserShellFolderLocation
 	$Desktop = @"
 "",
 "[.ShellClassInfo]",
-"LocalizedResourceName=@%SystemRoot%\System32\shell32.dll,-21769",
+"LocalizedResourceName=@shell32.dll,-21769",
 "IconResource=%SystemRoot%\System32\imageres.dll,-183"
 "@
 
 	$Documents = @"
 "",
 "[.ShellClassInfo]",
-"LocalizedResourceName=@%SystemRoot%\System32\shell32.dll,-21770",
+"LocalizedResourceName=@shell32.dll,-21770",
  "IconResource=%SystemRoot%\System32\imageres.dll,-112",
 "IconFile=%SystemRoot%\System32\shell32.dll",
 "IconIndex=-235"
@@ -6398,15 +6398,15 @@ function Set-UserShellFolderLocation
 	$Downloads = @"
 "",
 "[.ShellClassInfo]",
-"LocalizedResourceName=@%SystemRoot%\System32\shell32.dll,-21798",
+"LocalizedResourceName=@shell32.dll,-21798",
 "IconResource=%SystemRoot%\System32\imageres.dll,-184"
 "@
 
 	$Music = @"
 "",
 "[.ShellClassInfo]",
-"LocalizedResourceName=@%SystemRoot%\System32\shell32.dll,-21790",
-"InfoTip=@%SystemRoot%\System32\shell32.dll,-12689",
+"LocalizedResourceName=@shell32.dll,-21790",
+"InfoTip=@shell32.dll,-12689",
 "IconResource=%SystemRoot%\System32\imageres.dll,-108",
 "IconFile=%SystemRoot%\System32\shell32.dll","IconIndex=-237"
 "@
@@ -6414,8 +6414,8 @@ function Set-UserShellFolderLocation
 	$Pictures = @"
 "",
 "[.ShellClassInfo]",
-"LocalizedResourceName=@%SystemRoot%\System32\shell32.dll,-21779",
-"InfoTip=@%SystemRoot%\System32\shell32.dll,-12688",
+"LocalizedResourceName=@shell32.dll,-21779",
+"InfoTip=@shell32.dll,-12688",
 "IconResource=%SystemRoot%\System32\imageres.dll,-113",
 "IconFile=%SystemRoot%\System32\shell32.dll",
 "IconIndex=-236"
@@ -6424,8 +6424,8 @@ function Set-UserShellFolderLocation
 	$Videos = @"
 "",
 "[.ShellClassInfo]",
-"LocalizedResourceName=@%SystemRoot%\System32\shell32.dll,-21791",
-"InfoTip=@%SystemRoot%\System32\shell32.dll,-12690",
+"LocalizedResourceName=@shell32.dll,-21791",
+"InfoTip=@shell32.dll,-12690",
 "IconResource=%SystemRoot%\System32\imageres.dll,-189",
 "IconFile=%SystemRoot%\System32\shell32.dll",
 "IconIndex=-238"
@@ -6458,7 +6458,7 @@ function Set-UserShellFolderLocation
 
 			foreach ($UserFolder in @("Desktop", "Documents", "Downloads", "Music", "Pictures", "Videos"))
 			{
-				# Extract the localized user folders strings from shell32.dll
+				# Extract the localized user folders strings from %SystemRoot%\System32\shell32.dll
 				Write-Information -MessageData "" -InformationAction Continue
 				Write-Verbose -Message ($Localization.DriveSelect -f [WinAPI.GetStrings]::GetString($LocalizedUserFolderNameIDs[$UserFolder])) -Verbose
 
@@ -6492,7 +6492,7 @@ function Set-UserShellFolderLocation
 		{
 			foreach ($UserFolder in @("Desktop", "Documents", "Downloads", "Music", "Pictures", "Videos"))
 			{
-				# Extract the localized user folders strings from shell32.dll
+				# Extract the localized user folders strings from %SystemRoot%\System32\shell32.dll
 				Write-Information -MessageData "" -InformationAction Continue
 				Write-Verbose -Message ($Localization.UserFolderRequest -f [WinAPI.GetStrings]::GetString($LocalizedUserFolderNameIDs[$UserFolder])) -Verbose
 
@@ -6548,7 +6548,7 @@ function Set-UserShellFolderLocation
 		{
 			foreach ($UserFolder in @("Desktop", "Documents", "Downloads", "Music", "Pictures", "Videos"))
 			{
-				# Extract the localized user folders strings from shell32.dll
+				# Extract the localized user folders strings from %SystemRoot%\System32\shell32.dll
 				Write-Information -MessageData "" -InformationAction Continue
 				Write-Verbose -Message ($Localization.UserDefaultFolder -f [WinAPI.GetStrings]::GetString($LocalizedUserFolderNameIDs[$UserFolder])) -Verbose
 
@@ -7281,7 +7281,7 @@ function NetworkDiscovery
 	Path to an icon
 
 	.EXAMPLE
-	Set-Association -ProgramPath 'C:\SumatraPDF.exe' -Extension .pdf -Icon 'shell32.dll,100'
+	Set-Association -ProgramPath 'C:\SumatraPDF.exe' -Extension .pdf -Icon '%SystemRoot%\System32\shell32.dll,100'
 
 	.EXAMPLE
 	Set-Association -ProgramPath '%ProgramFiles%\Notepad++\notepad++.exe' -Extension .txt -Icon '%ProgramFiles%\Notepad++\notepad++.exe,0'
@@ -7562,7 +7562,7 @@ public static int UnloadHive(RegistryHives hive, string subKey)
 	[array]$Global:RegisteredProgIDs = @()
 
 	Write-Information -MessageData "" -InformationAction Continue
-	# Extract the localized "Please wait..." string from shell32.dll
+	# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 	Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 	# Register %1 argument if ProgId exists as an executable file
@@ -9049,7 +9049,7 @@ function Uninstall-UWPApps
 	$Window.Title               = $Localization.UWPAppsTitle
 	$ButtonUninstall.Content    = $Localization.Uninstall
 	$TextBlockRemoveForAll.Text = $Localization.UninstallUWPForAll
-	# Extract the localized "Select all" string from shell32.dll
+	# Extract the localized "Select all" string from %SystemRoot%\System32\shell32.dll
 	$TextBlockSelectAll.Text    = [WinAPI.GetStrings]::GetString(31276)
 
 	$ButtonUninstall.Add_Click({ButtonUninstallClick})
@@ -9071,7 +9071,7 @@ function Uninstall-UWPApps
 		)
 
 		Write-Information -MessageData "" -InformationAction Continue
-		# Extract the localized "Please wait..." string from shell32.dll
+		# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 		Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 		$AppxPackages = @(Get-AppxPackage -PackageTypeFilter Bundle -AllUsers:$AllUsers | Where-Object -FilterScript {$_.Name -notin $ExcludedAppxPackages})
@@ -9176,7 +9176,7 @@ function Uninstall-UWPApps
 	function ButtonUninstallClick
 	{
 		Write-Information -MessageData "" -InformationAction Continue
-		# Extract the localized "Please wait..." string from shell32.dll
+		# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 		Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
 		$Window.Close() | Out-Null
@@ -9322,8 +9322,9 @@ function Uninstall-UWPApps
 	Current user
 #>
 function Install-HEVC
+function Install-HEVC
 {
-	if (-not (Get-AppxPackage -Name Microsoft.Windows.Photos))
+	if (-not (Get-AppxPackage -Name Microsoft.Windows.Photos)) ### вытащить фотографии
 	{
 		Write-Information -MessageData "" -InformationAction Continue
 		Write-Verbose -Message ($Localization.PhotosNotInstalled, ($Localization.Skipped -f $MyInvocation.Line.Trim()) -join " ") -Verbose
@@ -9332,7 +9333,7 @@ function Install-HEVC
 		return
 	}
 
-	if (Get-AppxPackage -Name Microsoft.HEVCVideoExtension)
+	if (Get-AppxPackage -Name Microsoft.HEVCVideoExtension) ### вытащить HEVC
 	{
 		Write-Information -MessageData "" -InformationAction Continue
 		Write-Verbose -Message ($Localization.HEVCInstalled, ($Localization.Skipped -f $MyInvocation.Line.Trim()) -join " ") -Verbose
@@ -9342,19 +9343,19 @@ function Install-HEVC
 	}
 
 	Write-Information -MessageData "" -InformationAction Continue
-	# Extract the localized "Please wait..." string from shell32.dll
+	# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 	Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 
+	# Get actual HEVC package version
+	# https://github.com/farag2/Sophia-Script-for-Windows/blob/master/HEVC/HEVC_version.txt
 	try
 	{
-		$DownloadsFolder = Get-ItemPropertyValue -Path "HKCU:\Software\Microsoft\Windows\CurrentVersion\Explorer\User Shell Folders" -Name "{374DE290-123F-4565-9164-39C4925E467B}"
 		$Parameters = @{
-			Uri             = "https://raw.githubusercontent.com/farag2/Sophia-Script-for-Windows/refs/heads/master/HEVC/Microsoft.HEVCVideoExtension_8wekyb3d8bbwe.appx"
-			OutFile         = "$DownloadsFolder\Microsoft.HEVCVideoExtension_8wekyb3d8bbwe.appx"
+			Uri             = "https://raw.githubusercontent.com/farag2/Sophia-Script-for-Windows/refs/heads/master/HEVC/HEVC_version.txt"
 			UseBasicParsing = $true
 			Verbose         = $true
 		}
-		Invoke-WebRequest @Parameters
+		$HEVC_version = Invoke-WebRequest @Parameters
 	}
 	catch [System.Net.WebException]
 	{
@@ -9365,12 +9366,25 @@ function Install-HEVC
 		return
 	}
 
-	Write-Information -MessageData "" -InformationAction Continue
-	Write-Verbose -Message $Localization.HEVCInstallNotification -Verbose
-	Write-Information -MessageData "" -InformationAction Continue
+	if ([System.Version]$HEVC_version.Content -gt [System.Version](Get-AppxPackage -Name Microsoft.HEVCVideoExtension).Version
+	{
+		# Download HEVC
+		$DownloadsFolder = Get-ItemPropertyValue -Path "HKCU:\Software\Microsoft\Windows\CurrentVersion\Explorer\User Shell Folders" -Name "{374DE290-123F-4565-9164-39C4925E467B}"
+		$Parameters = @{
+			Uri             = "https://raw.githubusercontent.com/farag2/Sophia-Script-for-Windows/refs/heads/master/HEVC/Microsoft.HEVCVideoExtension_8wekyb3d8bbwe.appx"
+			OutFile         = "$DownloadsFolder\Microsoft.HEVCVideoExtension_8wekyb3d8bbwe.appx"
+			UseBasicParsing = $true
+			Verbose         = $true
+		}
+		Invoke-WebRequest @Parameters
 
-	Add-AppxPackage -Path "$DownloadsFolder\Microsoft.HEVCVideoExtension_8wekyb3d8bbwe.appx" -Verbose
-	Remove-Item -Path "$DownloadsFolder\Microsoft.HEVCVideoExtension_8wekyb3d8bbwe.appx" -Force
+		Write-Information -MessageData "" -InformationAction Continue
+		Write-Verbose -Message $Localization.HEVCInstallNotification -Verbose
+		Write-Information -MessageData "" -InformationAction Continue
+
+		Add-AppxPackage -Path "$DownloadsFolder\Microsoft.HEVCVideoExtension_8wekyb3d8bbwe.appx" -Verbose
+		Remove-Item -Path "$DownloadsFolder\Microsoft.HEVCVideoExtension_8wekyb3d8bbwe.appx" -Force
+	}
 }
 
 <#
@@ -11467,7 +11481,7 @@ function MSIExtractContext
 			$Value = "msiexec.exe /a `"%1`" /qb TARGETDIR=`"%1 extracted`""
 			New-ItemProperty -Path Registry::HKEY_CLASSES_ROOT\Msi.Package\shell\Extract\Command -Name "(default)" -PropertyType String -Value $Value -Force
 			New-ItemProperty -Path Registry::HKEY_CLASSES_ROOT\Msi.Package\shell\Extract -Name MUIVerb -PropertyType String -Value "@shell32.dll,-37514" -Force
-			New-ItemProperty -Path Registry::HKEY_CLASSES_ROOT\Msi.Package\shell\Extract -Name Icon -PropertyType String -Value "shell32.dll,-16817" -Force
+			New-ItemProperty -Path Registry::HKEY_CLASSES_ROOT\Msi.Package\shell\Extract -Name Icon -PropertyType String -Value "%SystemRoot%\System32\shell32.dll,-16817" -Force
 		}
 		"Hide"
 		{
@@ -12013,7 +12027,7 @@ function BitmapImageNewContext
 				try
 				{
 					Write-Information -MessageData "" -InformationAction Continue
-					# Extract the localized "Please wait..." string from shell32.dll
+					# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 					Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 					Write-Information -MessageData "" -InformationAction Continue
 
@@ -12093,7 +12107,7 @@ function RichTextDocumentNewContext
 				try
 				{
 					Write-Information -MessageData "" -InformationAction Continue
-					# Extract the localized "Please wait..." string from shell32.dll
+					# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 					Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 					Write-Information -MessageData "" -InformationAction Continue
 
@@ -12306,7 +12320,7 @@ function ScanRegistryPolicies
 	}
 
 	Write-Information -MessageData "" -InformationAction Continue
-	# Extract the localized "Please wait..." string from shell32.dll
+	# Extract the localized "Please wait..." string from %SystemRoot%\System32\shell32.dll
 	Write-Verbose -Message ([WinAPI.GetStrings]::GetString(12612)) -Verbose
 	Write-Information -MessageData "" -InformationAction Continue
